move test to WorkspaceRightBoxSection.test.tsx
