+++ src/components/RightBoxSection.test.tsx
@@ -1,10 +1,7 @@
 import { act, fireEvent, screen } from '@testing-library/react';
 import React from 'react';
 import { Ajax } from 'src/libs/ajax';
-import Events, { extractWorkspaceDetails } from 'src/libs/events';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
-import { defaultAzureWorkspace } from 'src/testing/workspace-fixtures';
-import { WorkspaceRightBoxSection } from 'src/workspaces/dashboard/WorkspaceRightBoxSection';
 
 import { RightBoxSection } from './RightBoxSection';
 
@@ -12,7 +9,6 @@ type AjaxContract = ReturnType<typeof Ajax>;
 jest.mock('src/libs/ajax');
 
 describe('RightBoxSection', () => {
-  const workspace = defaultAzureWorkspace;
   const captureEvent = jest.fn();
 
   beforeEach(() => {
@@ -49,33 +45,5 @@ describe('RightBoxSection', () => {
 
     // Assert
     expect(screen.getByText('Panel Content')).toBeInTheDocument();
-  });
-
-  it('fires a metrics event when the panel is toggled', async () => {
-    // Arrange
-    // Act
-    render(
-      <>
-        <WorkspaceRightBoxSection title='Test Title' persistenceId='metricsId' workspace={workspace} />
-        <RightBoxSection title='Test Title' persistenceId='metricsId' onOpenChangedCallback={captureEvent}>
-          Panel Content
-        </RightBoxSection>
-      </>
-    );
-    const titleElement = screen.getAllByText('Test Title');
-    fireEvent.click(titleElement[0]);
-    fireEvent.click(titleElement[0]);
-
-    // Assert
-    expect(captureEvent).toHaveBeenNthCalledWith(1, Events.workspaceDashboardToggleSection, {
-      opened: true,
-      title: 'Test Title',
-      ...extractWorkspaceDetails(workspace),
-    });
-    expect(captureEvent).toHaveBeenNthCalledWith(2, Events.workspaceDashboardToggleSection, {
-      opened: false,
-      title: 'Test Title',
-      ...extractWorkspaceDetails(workspace),
-    });
   });
 });
